requirements.txt
streamlit
google-generativeai
moviepy
pillow
numpy